Feature/indun2 (#790)
* [Game] WIP: Work with dungeons;
- code for working with dungeons and Indun tables for the database from Genesis;
- now each instance has its own Id for the team and for the personal dungeon;
- Added handling of events for players and NPCs when working with dungeons;
- Added check for the number of attempts to pass the dungeon;
- after death in the dungeon, the character is transferred to Priests of Nui into dungeon;

* - updated AI;
- updated DoodadFuncRemoveInstance;
- updated SetFaction;
- updated Buff.ScheduleEffect;
- calculation of distance without Z coordinate;

* instance_howling_abyss:
- updated NpcSpawnerIds;

* Unit Tests and Integration Tests updates and fixes

---------

+++ AAEmu.Game/Core/Packets/S2C/TCEmblemStreamRecvStatusPacket.cs
@@ -15,7 +15,7 @@ public class TCEmblemStreamRecvStatusPacket : StreamPacket
     public override PacketStream Write(PacketStream stream)
     {
         stream.Write((byte)_status); // status
-        stream.Write((int)0);
+        stream.Write(0);
 
         return stream;
     }
