+++ AAEmu.Game/Scripts/Commands/Nwrite.cs
@@ -87,7 +87,7 @@ public class Nwrite : ICommand
                             {
                                 if (doodad.Spawner.Id == 0) // spawned into the game manually
                                 {
-                                    var newId = (uint)((spawners[spawners.Count - 1].Id) + 1);
+                                    var newId = (spawners[spawners.Count - 1].Id) + 1;
                                     var pos = new JsonPosition();
                                     pos.X = doodad.Transform.World.Position.X;
                                     pos.Y = doodad.Transform.World.Position.Y;
@@ -167,7 +167,7 @@ public class Nwrite : ICommand
 
                                     if (npcs[i].Spawner.Id == 0) // spawned into the game manually
                                     {
-                                        var newId = (uint)((spawners[spawners.Count - 1].Id) + 1);
+                                        var newId = (spawners[spawners.Count - 1].Id) + 1;
 
                                         var pos = npcs[i].Transform.World;
 
@@ -246,7 +246,7 @@ public class Nwrite : ICommand
                             // Is it a new spawner ?
                             if (npc.Spawner.Id == 0) // spawned into the game manually
                             {
-                                var newId = (uint)((spawners[spawners.Count - 1].Id) + 1);
+                                var newId = (spawners[spawners.Count - 1].Id) + 1;
                                 var pos = npc.Transform.World;
 
                                 var newEntry = new JsonNpcSpawns();
